+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-#v1.1
+#Added
+
+- A complete new sound palette for handling the books
+- Added new textures